pyproject.toml
[project]
name = "three-tier-architecture"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "streamlit (>=1.51.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
